+++ text/src/text_render.rs
@@ -1,8 +1,8 @@
-use std::cell::RefCell;
+use std::{cell::RefCell, ops::Deref};
 
 use ribir_algo::Sc;
 use ribir_geom::{Rect, Size};
-use ribir_painter::{Painter, Path, PathStyle};
+use ribir_painter::{Painter, Path};
 
 use crate::{font_db::FontDB, GlyphBound, VisualGlyphs};
 
@@ -10,7 +10,7 @@ use crate::{font_db::FontDB, GlyphBound, VisualGlyphs};
 /// path style
 pub fn draw_glyphs_in_rect(
   painter: &mut Painter, visual_glyphs: VisualGlyphs, box_rect: Rect, font_size: f32,
-  path_style: &PathStyle, font_db: Sc<RefCell<FontDB>>,
+  font_db: Sc<RefCell<FontDB>>,
 ) {
   let visual_rect = visual_glyphs.visual_rect();
   let Some(paint_rect) = painter.intersection_paint_bounds(&box_rect) else {
@@ -20,19 +20,13 @@ pub fn draw_glyphs_in_rect(
     painter.clip(Path::rect(&paint_rect).into());
   }
   painter.translate(visual_rect.origin.x, visual_rect.origin.y);
-  draw_glyphs(
-    painter,
-    visual_glyphs.glyph_bounds_in_rect(&paint_rect),
-    font_size,
-    path_style,
-    font_db,
-  );
+  draw_glyphs(painter, visual_glyphs.glyph_bounds_in_rect(&paint_rect), font_size, font_db);
 }
 
 /// draw the glyphs with the given brush, font_size and path style
 pub fn draw_glyphs(
   painter: &mut Painter, glyphs: impl Iterator<Item = GlyphBound>, font_size: f32,
-  path_style: &PathStyle, font_db: Sc<RefCell<FontDB>>,
+  font_db: Sc<RefCell<FontDB>>,
 ) {
   glyphs.for_each(|g| {
     let font_db = font_db.borrow();
@@ -42,13 +36,15 @@ pub fn draw_glyphs(
       let unit = face.units_per_em() as f32;
       let scale = font_size / unit;
       let mut painter = painter.save_guard();
-      if let Some(path) = face.outline_glyph(g.glyph_id, path_style) {
+      if let Some(path) = face.outline_glyph(g.glyph_id) {
         painter
           .translate(g.bound.min_x(), g.bound.min_y())
           .scale(scale, -scale)
           .translate(0., -unit);
-
-        painter.fill_path(path.into());
+        match painter.style() {
+          ribir_painter::PathStyle::Fill => painter.fill_path(path.into()),
+          ribir_painter::PathStyle::Stroke => painter.stroke_path(path.deref().clone()),
+        };
       } else if let Some(svg) = face.glyph_svg_image(g.glyph_id) {
         let grid_scale = face
           .vertical_height()