+++ widgets/src/path.rs
@@ -1,60 +1,27 @@
 use ribir_core::prelude::*;
 
-/// Widget just use as a paint kit for a path and not care about its size. Use
-/// `[PathWidget]!` instead of.
+/// The widget serves as a painting kit for a path and does not concern itself
+/// with size. Otherwise, it should use `Resource<Path>` instead.
 #[derive(Declare, Clone)]
 pub struct PathPaintKit {
-  pub path: Path,
-  #[declare(default)]
-  pub style: PathStyle,
-}
-
-macro_rules! paint_method {
-  () => {
-    fn paint(&self, ctx: &mut PaintingCtx) {
-      let painter = ctx.painter();
-      match &self.style {
-        PathStyle::Fill => painter.fill_path(self.path.clone().into()),
-        PathStyle::Stroke(strokes) => painter
-          .set_strokes(strokes.clone())
-          .stroke_path(self.path.clone().into()),
-      };
-    }
-  };
+  pub path: Resource<Path>,
 }
 
 impl Render for PathPaintKit {
   #[inline]
-  fn perform_layout(&self, clamp: BoxClamp, _: &mut LayoutCtx) -> Size { clamp.max }
+  fn perform_layout(&self, _: BoxClamp, _: &mut LayoutCtx) -> Size { Size::zero() }
 
   #[inline]
   fn only_sized_by_parent(&self) -> bool { true }
 
-  paint_method!();
+  fn paint(&self, ctx: &mut PaintingCtx) {
+    let path = PaintPath::Share(self.path.clone());
+    ctx.painter().draw_path(path);
+  }
 
   fn hit_test(&self, _ctx: &HitTestCtx, _: Point) -> HitTest {
     HitTest { hit: false, can_hit_child: false }
   }
-}
-
-#[derive(Declare)]
-/// A path widget which size careful and can process events only if user hit at
-/// the path self, not its size cover area.
-pub struct PathWidget {
-  pub path: Path,
-  #[declare(default)]
-  pub style: PathStyle,
-}
-
-/// Path widget just use as a paint kit for a path and not care about its size.
-/// Use `[HitTesPath]!` instead of.
-impl Render for PathWidget {
-  #[inline]
-  fn perform_layout(&self, _: BoxClamp, _: &mut LayoutCtx) -> Size {
-    self.path.bounds().max().to_vector().to_size()
-  }
-
-  paint_method!();
 }
 
 #[cfg(test)]
@@ -64,8 +31,10 @@ mod tests {
 
   use super::*;
 
-  fn circle40() -> Path { Path::circle(Point::new(20., 20.), 20.) }
+  fn circle40() -> Resource<Path> { Path::circle(Point::new(20., 20.), 20.).into() }
   const WND_SIZE: Size = Size::new(48., 48.);
+  const SIZE_40: Size = Size::new(40., 40.);
+
   widget_test_suit!(
     circle40_kit,
     WidgetTester::new(fn_widget! {
@@ -76,19 +45,45 @@ mod tests {
     })
     .with_wnd_size(WND_SIZE)
     .with_comparison(0.000025),
-    LayoutCase::default().with_size(WND_SIZE)
+    LayoutCase::default().with_size(Size::zero())
   );
 
   widget_test_suit!(
     circle40,
     WidgetTester::new(fn_widget! {
-      @PathWidget {
-        path: circle40(),
+      let path = circle40();
+      @ $path { foreground: Color::BLACK }
+    })
+    .with_wnd_size(WND_SIZE)
+    .with_comparison(0.000025),
+    LayoutCase::default().with_size(SIZE_40)
+  );
+
+  widget_test_suit!(
+    fill_circle40,
+    WidgetTester::new(fn_widget! {
+      let path = circle40();
+      @ $path {
+        painting_style: PaintingStyle::Fill,
         foreground: Color::BLACK,
       }
     })
     .with_wnd_size(WND_SIZE)
     .with_comparison(0.000025),
-    LayoutCase::default().with_size(Size::new(40., 40.))
+    LayoutCase::default().with_size(SIZE_40)
+  );
+
+  widget_test_suit!(
+    stroke_circle40,
+    WidgetTester::new(fn_widget! {
+      let path = circle40();
+      @ $path {
+        painting_style: PaintingStyle::Stroke(StrokeOptions::default()),
+        foreground: Color::BLACK,
+      }
+    })
+    .with_wnd_size(WND_SIZE)
+    .with_comparison(0.000025),
+    LayoutCase::default().with_size(SIZE_40)
   );
 }